+++ extensions/xcode/src/user-interfaces/xcode-create-swift-playground/xcode-create-swift-playground-form.user-interface.tsx
@@ -1,20 +1,10 @@
-import {
-  ActionPanel,
-  Form,
-  getPreferenceValues,
-  Navigation,
-  showHUD,
-  showToast,
-  SubmitFormAction,
-  ToastStyle,
-} from "@raycast/api";
+import { Action, ActionPanel, Form, Navigation, showHUD, showToast, Toast } from "@raycast/api";
 import { XcodeSwiftPlaygroundService } from "../../services/xcode-swift-playground.service";
 import { XcodeSwiftPlaygroundPlatform } from "../../models/swift-playground/xcode-swift-playground-platform.model";
 import { XcodeSwiftPlayground } from "../../models/swift-playground/xcode-swift-playground.model";
 import { XcodeSwiftPlaygroundCreationParameters } from "../../models/swift-playground/xcode-swift-playground-creation-parameters.model";
 import { XcodeSwiftPlaygroundTemplate } from "../../models/swift-playground/xcode-swift-playground-template.model";
 import tildify from "tildify";
-import { Preferences } from "../../models/utils/preferences.model";
 
 /**
  * Xcode create Swift Playground Form
@@ -25,28 +15,31 @@ export function xcodeCreateSwiftPlaygroundForm(
   xcodeSwiftPlaygroundService: XcodeSwiftPlaygroundService,
   navigation: Navigation
 ): JSX.Element {
-  const preferences = getPreferenceValues<Preferences>();
   return (
     <Form
       actions={
         <ActionPanel>
-          <SubmitFormAction
+          <Action.SubmitForm
             title={"Open or create Swift Playground"}
             onSubmit={(formValues) => {
-              onFormSubmit(formValues, xcodeSwiftPlaygroundService, navigation, false);
+              return onFormSubmit(formValues, xcodeSwiftPlaygroundService, navigation, false);
             }}
           />
-          <SubmitFormAction
+          <Action.SubmitForm
             title={"Create Swift Playground"}
             onSubmit={(formValues) => {
-              onFormSubmit(formValues, xcodeSwiftPlaygroundService, navigation, true);
+              return onFormSubmit(formValues, xcodeSwiftPlaygroundService, navigation, true);
             }}
           />
         </ActionPanel>
       }
     >
       <Form.TextField id="name" title={"Name"} defaultValue="MyPlayground" />
-      <Form.TextField id="location" title="Location" defaultValue={preferences.playgroundDefaultLocation} />
+      <Form.TextField
+        id="location"
+        title="Location"
+        defaultValue={xcodeSwiftPlaygroundService.defaultSwiftPlaygroundLocation}
+      />
       <Form.Dropdown id="platform" title="Platform" defaultValue={XcodeSwiftPlaygroundPlatform.iOS}>
         {Object.keys(XcodeSwiftPlaygroundPlatform)
           .map((platform) => platform.toLocaleLowerCase())
@@ -54,15 +47,9 @@ export function xcodeCreateSwiftPlaygroundForm(
             return <Form.Dropdown.Item key={platform} value={platform} title={platform.replace("os", "OS")} />;
           })}
       </Form.Dropdown>
-      <Form.Dropdown id="template" title="Template" defaultValue={XcodeSwiftPlaygroundTemplate.empty}>
-        {Object.keys(XcodeSwiftPlaygroundTemplate).map((platform) => {
-          return (
-            <Form.Dropdown.Item
-              key={platform}
-              value={platform}
-              title={platform.charAt(0).toUpperCase() + platform.slice(1)}
-            />
-          );
+      <Form.Dropdown id="template" title="Template" defaultValue={XcodeSwiftPlaygroundTemplate.Empty}>
+        {Object.keys(XcodeSwiftPlaygroundTemplate).map((template) => {
+          return <Form.Dropdown.Item key={template} value={template} title={template} />;
         })}
       </Form.Dropdown>
       <Form.Checkbox id="open" label="Open in Xcode after creation" defaultValue={true} />
@@ -75,6 +62,7 @@ export function xcodeCreateSwiftPlaygroundForm(
  * @param formValues The Form values
  * @param xcodeSwiftPlaygroundService XcodeSwiftPlaygroundService
  * @param navigation The Navigation
+ * @param forceCreate Bool value if creation of a Swift Playground should be enforced
  */
 async function onFormSubmit(
   formValues: any,
@@ -94,14 +82,20 @@ async function onFormSubmit(
     // Log error
     console.log(error);
     // Show failure Toast
-    await showToast(ToastStyle.Failure, "An error occurred while creating the Swift Playground");
+    await showToast({
+      style: Toast.Style.Failure,
+      title: "An error occurred while creating the Swift Playground",
+    });
     // Return out of function
     return;
   }
   // Check if Swift Playground already exists and should not be opened
   if (swiftPlayground.alreadyExists && !formValues.open) {
     // Inform user that the Playground already exists
-    return showToast(ToastStyle.Failure, "Swift Playground already exists");
+    return showToast({
+      style: Toast.Style.Failure,
+      title: "Swift Playground already exists",
+    });
   }
   // Initialize success message title
   const successMessageTitle = [
@@ -110,7 +104,7 @@ async function onFormSubmit(
     "at",
     tildify(swiftPlayground.path),
   ].join(" ");
-  // Check if should open Swift Playground after creation
+  // Check if Swift Playground should be opened after creation
   if (formValues.open) {
     try {
       // Open Swift Playground
@@ -119,11 +113,17 @@ async function onFormSubmit(
       await showHUD(successMessageTitle);
     } catch {
       // Show failure Toast
-      await showToast(ToastStyle.Failure, "Swift Playground could not be opened");
+      await showToast({
+        style: Toast.Style.Failure,
+        title: "Swift Playground could not be opened",
+      });
     }
   } else {
     // Show success Toast
-    await showToast(ToastStyle.Success, successMessageTitle);
+    await showToast({
+      style: Toast.Style.Success,
+      title: successMessageTitle,
+    });
   }
   // Pop to root
   navigation.pop();