+++ extensions/common-directory/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Common Directory Changelog
 
+## [Added Common Directory] - 2022-04-30
+
+- Fix Rank algorithm bugs
+- Add hazardous operation warning
+
 ## [Added Common Directory] - 2022-04-25
-- Release of Common Directory version 1.0
+
 - Quick access to frequently used folder directories